# Conflicts:
#	.gitignore
#	.vscode/tasks.json
#	i21Database/dbo/Stored Procedures/uspGRReportGrainReceiptLSM.sql
#	i21Database/dbo/Tables/tblAREACustomerExport.sql
#	i21Database/dbo/Tables/tblGRCompanyPreference.sql
#	i21Database/dbo/Tables/tblSCDeliverySheet.sql
#	i21Database/dbo/Views/vyuAPREceiptClearing.sql
#	i21Database/dbo/Views/vyuGRTransferChargesClearing.sql
#	i21Database/dbo/Views/vyuGRTransferClearing.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/Tables/tblIPStorageLocationError.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE tblIPStorageLocationError (
 	intStorageLocationErrorId INT identity(1, 1)
 	,intStorageLocationStageId INT
-	,intTrxSequenceNo INT
+	,intTrxSequenceNo BIGINT
 	,strCompanyLocation NVARCHAR(6) COLLATE Latin1_General_CI_AS
 	,intActionId INT
 	,dtmCreatedDate DATETIME
